Hago andar test sin terminar y sin fetch lazy

+++ src/main/java/ar/edu/unq/epers/bichomon/backend/dao/BichoDAO.java
@@ -6,5 +6,5 @@ public interface BichoDAO {
 
     void guardar(Bicho bicho);
 
-    Bicho recuperar(Long id);
+    Bicho recuperar(Integer id);
 }
